Refactored validator package due growing issues
