+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/HearingBooking.java
@@ -4,9 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.fpl.interfaces.HearingBookingDetailsGroup;
-import uk.gov.hmcts.reform.fpl.validators.interfaces.HasFutureDate;
-import uk.gov.hmcts.reform.fpl.validators.interfaces.HasStartDateAfterEndDate;
-import uk.gov.hmcts.reform.fpl.validators.interfaces.TimeNotZero;
+import uk.gov.hmcts.reform.fpl.validation.interfaces.HasFutureDate;
+import uk.gov.hmcts.reform.fpl.validation.interfaces.HasStartDateAfterEndDate;
+import uk.gov.hmcts.reform.fpl.validation.interfaces.TimeNotZero;
 
 import java.time.LocalDateTime;
 import java.util.List;
