+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/common/Document.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.fpl.interfaces.UploadDocumentsGroup;
-import uk.gov.hmcts.reform.fpl.validators.interfaces.HasAttachedDocument;
+import uk.gov.hmcts.reform.fpl.validation.interfaces.HasAttachedDocument;
 
 import javax.validation.constraints.NotBlank;
