+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/interfaces/HasContactDirection.java
@@ -1,19 +1,20 @@
-package uk.gov.hmcts.reform.fpl.validators.interfaces;
+package uk.gov.hmcts.reform.fpl.validation.interfaces;
 
-import uk.gov.hmcts.reform.fpl.validators.HasDocumentsIncludedInSwetValidator;
+import uk.gov.hmcts.reform.fpl.validation.validators.HasContactDirectionValidator;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
 import javax.validation.Constraint;
 import javax.validation.Payload;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { HasDocumentsIncludedInSwetValidator.class })
-public @interface HasDocumentsIncludedInSwet {
-    String message() default "Attach the SWET or change the status from 'Included in SWET'.";
+@Constraint(validatedBy = { HasContactDirectionValidator.class })
+public @interface HasContactDirection {
+    String message() default "Enter the contact's full name";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
 }
