+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/interfaces/HasDocumentsIncludedInSwet.java
@@ -1,6 +1,6 @@
-package uk.gov.hmcts.reform.fpl.validators.interfaces;
+package uk.gov.hmcts.reform.fpl.validation.interfaces;
 
-import uk.gov.hmcts.reform.fpl.validators.HasAttachedDocumentValidator;
+import uk.gov.hmcts.reform.fpl.validation.validators.documents.HasDocumentsIncludedInSwetValidator;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -11,9 +11,9 @@ import javax.validation.Payload;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { HasAttachedDocumentValidator.class })
-public @interface HasAttachedDocument {
-    String message() default "Attach the document or change the status from 'Attached'.";
+@Constraint(validatedBy = { HasDocumentsIncludedInSwetValidator.class })
+public @interface HasDocumentsIncludedInSwet {
+    String message() default "Attach the SWET or change the status from 'Included in SWET'.";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
 }
