+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/interfaces/HasTelephoneOrMobile.java
@@ -1,19 +1,20 @@
-package uk.gov.hmcts.reform.fpl.validators.interfaces;
+package uk.gov.hmcts.reform.fpl.validation.interfaces;
 
-import uk.gov.hmcts.reform.fpl.validators.HasFutureDateValidator;
+import uk.gov.hmcts.reform.fpl.validation.validators.HasTelephoneOrMobileValidator;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
 import javax.validation.Constraint;
 import javax.validation.Payload;
 
-@Target(ElementType.FIELD)
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = HasFutureDateValidator.class)
-public @interface HasFutureDate {
-    String message() default "Enter a date in the future";
+@Constraint(validatedBy = { HasTelephoneOrMobileValidator.class })
+public @interface HasTelephoneOrMobile {
+    String message() default "Enter at least one telephone number for the contact";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
 }
