+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/validators/HasContactDirectionValidator.java
@@ -1,8 +1,8 @@
-package uk.gov.hmcts.reform.fpl.validators;
+package uk.gov.hmcts.reform.fpl.validation.validators;
 
 import org.apache.commons.lang.StringUtils;
 import uk.gov.hmcts.reform.fpl.model.interfaces.TelephoneContacts;
-import uk.gov.hmcts.reform.fpl.validators.interfaces.HasContactDirection;
+import uk.gov.hmcts.reform.fpl.validation.interfaces.HasContactDirection;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
