+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/validators/HasTelephoneOrMobileValidator.java
@@ -1,8 +1,8 @@
-package uk.gov.hmcts.reform.fpl.validators;
+package uk.gov.hmcts.reform.fpl.validation.validators;
 
 import org.apache.commons.lang.StringUtils;
 import uk.gov.hmcts.reform.fpl.model.interfaces.TelephoneContacts;
-import uk.gov.hmcts.reform.fpl.validators.interfaces.HasTelephoneOrMobile;
+import uk.gov.hmcts.reform.fpl.validation.interfaces.HasTelephoneOrMobile;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
