+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/validators/time/HasFutureDateValidator.java
@@ -1,6 +1,6 @@
-package uk.gov.hmcts.reform.fpl.validators;
+package uk.gov.hmcts.reform.fpl.validation.validators.time;
 
-import uk.gov.hmcts.reform.fpl.validators.interfaces.HasFutureDate;
+import uk.gov.hmcts.reform.fpl.validation.interfaces.HasFutureDate;
 
 import java.time.LocalDateTime;
 import javax.validation.ConstraintValidator;
