+++ service/src/test/java/uk/gov/hmcts/reform/fpl/validation/HasAttachedDocumentValidatorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.fpl.validators;
+package uk.gov.hmcts.reform.fpl.validation;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
